g
g

+++ app/src/main/java/com/itzik/hotelapp/ui/theme/project/model/User.kt
@@ -16,6 +16,6 @@ data class User(
     val password: String,
     var isLoggedIn: Boolean = false,
     var isItemLiked: Boolean,
-    var phoneNumber: Int,
+    var phoneNumber: Long,
     var profileImage: String? = null
 ) : Parcelable
